+++ worlds/banjo_tooie/__init__.py
@@ -181,11 +181,13 @@ class BanjoTooieWorld(World):
                 i == (self.options.extra_trebleclefs_count.value*3 - 15):
                     break
                 trap_big_pants_counter += 1
-        if self.options.traps_nests_ratio.value > 0:
-            total_nests = 315 + 135
+        if self.options.traps.value == 1 and self.options.nestsanity.value == True:
+            total_nests = all_item_table[itemName.ENEST].qty + all_item_table[itemName.FNEST].qty - 23
             removed_nests = int(total_nests * self.options.traps_nests_ratio.value / 100)
-            removed_enests = int(315 * self.options.traps_nests_ratio.value / 100)
-            removed_fnests = removed_nests - removed_enests
+            removed_enests = int((all_item_table[itemName.ENEST].qty-16) * self.options.traps_nests_ratio.value / 100)
+            removed_fnests = removed_nests - (removed_enests)
+            removed_enests += 16 #golden eggs
+            removed_fnests += 7 # golden eggs
             trap_big_pants_counter += removed_nests
         if self.options.traps.value == True:
             trup = divmod(trap_big_pants_counter, 4)
@@ -227,14 +229,18 @@ class BanjoTooieWorld(World):
                     elif item.code == 1230789:
                         for i in range(sqtrap_qty):
                             itempool += [self.create_item(name)]
+
                     #end of none qty logic
                     
                     #nests removal for nestsanity and nest traps
-                    elif item.code == 1230806 and self.options.traps_nests_ratio.value > 0:
-                        for i in range(315 - removed_enests):
+                    elif item.code == 1230806:
+                        for i in range(all_item_table[item.name].qty - removed_enests):
                             itempool += [self.create_item(name)]
-                    elif item.code == 1230807 and self.options.traps_nests_ratio.value > 0:
-                        for i in range(135 - removed_fnests):
+                    elif item.code == 1230807:
+                        for i in range(all_item_table[item.name].qty - removed_fnests):
+                            itempool += [self.create_item(name)]
+                    elif item.code == 1230805 and self.options.traps.value == True:
+                        for i in range(23):
                             itempool += [self.create_item(name)]
                     #end of nest removal for nest traps
                     
@@ -443,8 +449,6 @@ class BanjoTooieWorld(World):
             raise ValueError("You cannot have progressive bash attack without randomizing Stop N Swap and randomizing BK moves")
         if self.options.randomize_moves == False and self.options.jamjars_silo_costs.value != 0:
             raise ValueError("You cannot change the silo costs without randomizing Jamjars' moves.")
-        if self.options.nestsanity.value == False and self.options.traps_nests_ratio.value > 0:
-            raise ValueError("You cannot have nests as traps without nestsanity.")
         if self.options.open_hag1.value == False and self.options.victory_condition.value == 4:
             self.options.open_hag1.value = True
         if self.options.egg_behaviour.value == 1:
@@ -828,7 +832,6 @@ class BanjoTooieWorld(World):
             else:
                 return level
 
-        
         if self.options.randomize_world_loading_zone.value == False:
             return